sfil_jpeg.c: Fix an off-by-one with JPEG scanlines.
The decoder was adding 'stride' before decoding anything, which
resulted in the first line of the texture being garbage, and it
ended up writing past the end of the texture. In most cases, since
textures are a power of two, writing past the end didn't do anything,
but if the JPEG's height is a power of two, this could result in a
crash.